requirements.txt
Flask
psycopg2-binary
python-dotenv==0.19.2